remove Development and Production from checks namespace
All checks are now in the same namespace `BeyondCode\SelfDiagnosis\Checks` because the environment they are run in is now configurable.
